# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RW Digital API",
        "description": "RW Digital API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/warga": {
            "get": {
                "tags": [
                    "Warga"
                ],
                "summary": "List of Warga",
                "operationId": "d0da63bc0505db195a84cafa949b8d5d",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "keyword"
                    },
                    {
                        "name": "filter[keyword]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "keyword"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "name"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Warga"
                ],
                "summary": "Store Warga",
                "operationId": "3c7045f9972c4ceead481380483fbca8",
                "requestBody": {
                    "description": "Body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nik": {
                                        "$ref": "#/components/schemas/Warga/properties/nik"
                                    },
                                    "alamat_ktp": {
                                        "$ref": "#/components/schemas/Warga/properties/alamat_ktp"
                                    },
                                    "blok": {
                                        "$ref": "#/components/schemas/Warga/properties/blok"
                                    },
                                    "nomor": {
                                        "$ref": "#/components/schemas/Warga/properties/nomor"
                                    },
                                    "rt": {
                                        "$ref": "#/components/schemas/Warga/properties/rt"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data sukses disimpan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nik": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "nik field is required."
                                                ]
                                            }
                                        },
                                        "alamat_ktp": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "alamat_ktp field is required."
                                                ]
                                            }
                                        },
                                        "blok": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "blok field is required."
                                                ]
                                            }
                                        },
                                        "nomor": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "nomor field is required."
                                                ]
                                            }
                                        },
                                        "rt": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "rt field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/warga/import": {
            "post": {
                "tags": [
                    "Warga"
                ],
                "summary": "Import data from Excel to SQL table",
                "operationId": "36044c52891eb90d60a83fa5ee1708b4",
                "parameters": [
                    {
                        "name": "audit",
                        "in": "path",
                        "description": "Audit ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Excel file for data import",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file_template": {
                                        "description": "Excel file for data import",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "is_imported": {
                                        "description": "Set to TRUE to indicate that the data is imported",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK. The data import and storage were successful."
                    },
                    "401": {
                        "description": "Unauthorized. The 'is_imported' parameter must be set to TRUE for using this resource."
                    },
                    "422": {
                        "description": "Unprocessable Entity. Invalid input or data could not be processed."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/warga/{no_kk}/anggota": {
            "get": {
                "tags": [
                    "Warga"
                ],
                "summary": "Anggota warga details",
                "operationId": "7ddb3bb024022a7172a0c75b24545902",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Warga ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        }
    },
...
}